+++ carpa/urls.py
@@ -21,6 +21,8 @@ from django.conf.urls.static import static
 
 urlpatterns = [
     path('', include('main.urls')),
+    path('', include('reports.urls')),
+    path('', include('harvests.urls')),
     path('admin/', admin.site.urls),
 ]
