+++ main/views.py
@@ -1,6 +1,4 @@
-from django.http import HttpResponse
 from django.shortcuts import render
-
 
 
 def home(request):
@@ -9,20 +7,3 @@ def home(request):
 def login(request):   
     return render(request, "login.html")
 
-def people(request):   
-    return render(request, "people.html")
-
-def vehicles(request):   
-    return render(request, "vehicles.html")
-
-def origin(request):   
-    return render(request, "origin.html")
-
-def reports(request):   
-    return render(request, "reports.html")
-
-def create_report(request):   
-    return render(request, "create_report.html")
-
-def query_report(request):   
-    return render(request, "query_report.html")
