+++ reports/templates/query_report.html
